+++ Assets/scripts/Entity/Player/PlayerLevel.cs
@@ -58,10 +58,10 @@ namespace CrystalFlux.EntitySystem
 
             if (TryGetComponent<ISkillPointHolder>(out var sph)) sph.AddSkillPoints(1);
 
-            esm.AddStat(new(StatType.maxHp, 3));
-            esm.AddStat(new(StatType.attack, 1));
-            esm.AddStat(new(StatType.Intelligence, 1));
-            esm.AddStat(new(StatType.moveSpeed, 0.005f));
+            esm.AddStat(new(StatType.maxHp, 5));
+            esm.AddStat(new(StatType.attack, 2));
+            esm.AddStat(new(StatType.Intelligence, 2));
+            esm.AddStat(new(StatType.moveSpeed, 0.008f));
 
             if (pum != null) pum.TriggerUpgrades(PlayerUpgrade.TriggerCondition.OnLevelUp);