+++ Assets/scripts/Projectile/IChargeRegister.cs
@@ -1,0 +1,8 @@
+namespace CrystalFlux.ProjectileSystem
+{
+    public interface IChargeRegister
+    {
+        void RegisterChargedProjectile(Projectile p);
+        void UnregisterChargedProjectile(Projectile p);
+    }
+}